+++ parameters.tf
@@ -515,6 +515,15 @@ resource "aws_ssm_parameter" "issue_proxy_url" {
   tags      = module.labels.tags
 }
 
+resource "aws_ssm_parameter" "security_allow_no_token" {
+  count     = contains(var.optional_parameters_to_include, "security_allow_no_token") ? 1 : 0
+  overwrite = true
+  name      = format("%ssecurity_allow_no_token", local.config_var_prefix)
+  type      = "String"
+  value     = var.security_allow_no_token
+  tags      = module.labels.tags
+}
+
 resource "aws_ssm_parameter" "security_callback_rate_limit_request_count" {
   count     = contains(var.optional_parameters_to_include, "security_callback_rate_limit_request_count") ? 1 : 0
   overwrite = true