+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 
 ## Unreleased
-
+- Added: Optional parameters "security_allow_no_token"
 
 ## [0.1.16] 2020-12-14
 - Added: Missing variables for proxy URLs